Click playback slider to scrub (#32)
Added support for scrubbing / seeking by clicking the playback slider. (This change replaces the progress bar with an NSSlider to better support user interaction.) The slider handles continuous changes and stops updating in response to mpv position updates during click-and-drag scrubbing. This also adds a seeking property to media player, currently used to avoid sending a second seek command to mpv while waiting for the first one to be handled.